+++ src/components/GuideLinks.tsx
@@ -1,0 +1,5 @@
+interface GuideLinksProps {}
+
+export function GuideLinks(props: GuideLinksProps) {
+    return null;
+}
